RavenDB-5854 - fix failing tests - changed the total installed memory to double - added the option to provide the license inside a file - allow to skip logging lease license errors

+++ src/Raven.Server/Commercial/DetailsPerNode.cs
@@ -8,7 +8,7 @@ namespace Raven.Server.Commercial
 
         public int NumberOfCores;
 
-        public long InstalledMemoryInGb;
+        public double InstalledMemoryInGb;
 
         public double UsableMemoryInGb;
